worked on the comment
